// src/types/teams.ts
export type Team =
  | "electronics"
  | "dev-ops"
  | "web-team"
  | "ai-research"
  | "cloud-computing"
  | "cyber-security"
  | "data-science"
  | "game-development"
  | "iot"
  | "mobile-development"
  | "blockchain"
  | "networking"
  | "robotics"
  | "software-engineering"
  | "system-administration"
  | "ui-ux-design"
  | "database-management"
  | "bioinformatics"
  | "embedded-systems"
  | "quantum-computing";